refactor break app in-part
