+++ botstory/matchers.py
@@ -15,6 +15,27 @@ def build_deserialize(cls):
     return default_deserialize
 
 
+def get_validator(receive):
+    """
+    ask every matcher whether it can serve such filter
+
+    :param receive:
+    :return:
+    """
+    for matcher_type, m in matchers.items():
+        if hasattr(m, 'can_handle') and m.can_handle(receive):
+            receive = m.handle(receive)
+
+    return receive
+    # TODO: should
+    # if isinstance(receive, list):
+    #     return any.AnyOf([get_validator(r) for r in receive])
+    # elif is_string(receive):
+    #     return text.Match(receive)
+    # else:
+    #     return receive
+
+
 def matcher():
     def register(m):
         m.type = getattr(m, 'type', m.__name__)
