+++ botstory/middlewares/any/any.py
@@ -10,7 +10,6 @@ class Any:
         return True
 
 
-
 @matchers.matcher()
 class AnyOf:
     def __init__(self, list_of_matchers):
@@ -27,3 +26,11 @@ class AnyOf:
     @staticmethod
     def deserialize(data):
         return AnyOf([matchers.deserialize(d) for d in data])
+
+    @staticmethod
+    def can_handle(data):
+        return isinstance(data, list)
+
+    @staticmethod
+    def handle(data):
+        return AnyOf([matchers.get_validator(r) for r in data])
