+++ botstory/middlewares/location/test_location.py
@@ -5,7 +5,7 @@ from ...utils import answer, build_fake_session, build_fake_user, SimpleTrigger
 
 def teardown_function(function):
     print('tear down!')
-    story.core.clear()
+    story.core_instance.clear()
 
 
 def test_should_trigger_on_any_location():
